Extract and significantly expand the reproduction scope
