+++ e2e/test/scenarios/filters-reproductions/dashboard-filters-reproductions.cy.spec.js
@@ -1,6 +1,5 @@
 import moment from "moment-timezone"; // eslint-disable-line no-restricted-imports -- deprecated usage
 
-import { SAMPLE_DB_ID } from "e2e/support/cypress_data";
 import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
 import {
   ADMIN_PERSONAL_COLLECTION_ID,
@@ -39,7 +38,6 @@ import {
   visitPublicDashboard,
   setModelMetadata,
   tableHeaderClick,
-  questionInfoButton,
 } from "e2e/support/helpers";
 import {
   createMockDashboardCard,
@@ -3422,181 +3420,6 @@ describe("issue 35852", () => {
       });
     });
   }
-});
-
-describe("issue 35954", () => {
-  const questionDetails = {
-    name: "35954",
-    query: {
-      "source-table": REVIEWS_ID,
-      limit: 2,
-    },
-  };
-
-  const dashboardDetails = {
-    name: "35954D",
-  };
-
-  const updatedQuestionDetails = {
-    dataset_query: {
-      database: SAMPLE_DB_ID,
-      type: "native",
-      native: {
-        "template-tags": {
-          RATING: {
-            type: "dimension",
-            name: "RATING",
-            id: "017b9185-c7cc-41ec-ba17-b8b21af879cc",
-            "display-name": "SQL Filter",
-            default: null,
-            dimension: ["field", REVIEWS.RATING, null],
-            "widget-type": "number/=",
-            options: null,
-          },
-        },
-        collection: "REVIEWS",
-        query: "SELECT * FROM REVIEWS WHERE {{RATING}} LIMIT 2",
-      },
-    },
-    parameters: [
-      {
-        id: "017b9185-c7cc-41ec-ba17-b8b21af879cc",
-        type: "number/=",
-        target: ["dimension", ["template-tag", "RATING"]],
-        slug: "RATING",
-      },
-    ],
-    parameter_mappings: [],
-  };
-
-  beforeEach(() => {
-    restore();
-    cy.signInAsAdmin();
-
-    cy.createQuestionAndDashboard({
-      questionDetails,
-      cardDetails: {
-        size_x: 16,
-        size_y: 8,
-      },
-      dashboardDetails,
-    }).then(({ body: { dashboard_id, card_id } }) => {
-      cy.wrap(card_id).as("questionId");
-      cy.wrap(dashboard_id).as("dashboardId");
-
-      cy.request("PUT", `/api/card/${card_id}`, updatedQuestionDetails);
-
-      visitDashboard(dashboard_id);
-    });
-  });
-
-  it("should not break the UI when the dashboard filter loses connection (metabase#35954)", () => {
-    editDashboard();
-    cy.log("Add the number filter");
-    setFilter("Number");
-    connectFilterToColumn("SQL Filter");
-    saveDashboard();
-
-    cy.log("Give it a value and make sure that the filer applies");
-    filterWidget().click();
-    popover().within(() => {
-      cy.findByPlaceholderText("Enter a number").type(3).blur();
-      cy.button("Add filter").click();
-    });
-    cy.findAllByTestId("cell-data")
-      .should("contain", "kale")
-      .and("contain", "pete")
-      .and("not.contain", "xavier");
-
-    cy.log("Drill down to the question from the dashboard");
-    cy.findByTestId("legend-caption-title").click();
-    cy.get("@questionId").then(id => {
-      cy.location("pathname").should(
-        "eq",
-        `/question/${id}-${questionDetails.name}`,
-      );
-      cy.location("search").should("eq", "?RATING=3");
-    });
-
-    cy.log("Revert the question to its original (GUI) version");
-    cy.intercept("POST", "/api/revision/revert").as("revertQuestion");
-    questionInfoButton().click();
-    cy.findByTestId("saved-question-history-list")
-      .find("li")
-      .filter(":contains(You created this)")
-      .findByTestId("question-revert-button")
-      .click();
-    cy.wait("@revertQuestion");
-    // Mid-test assertions to root out the flakiness
-    cy.findByTestId("saved-question-history-list").should(
-      "contain",
-      "You edited this",
-    );
-    cy.findByTestId("saved-question-history-list")
-      .findAllByTestId("question-revert-button")
-      .should("have.length", 2);
-
-    cy.findByLabelText(`Back to ${dashboardDetails.name}`).click();
-
-    cy.get("@dashboardId").then(id => {
-      cy.location("pathname").should(
-        "eq",
-        `/dashboard/${id}-${dashboardDetails.name.toLowerCase()}`,
-      );
-    });
-
-    cy.log("Make sure the disconnected filter doesn't break UI");
-    // The filter widget will still have the number 3 applied as the filter,
-    // but that shouldn't affect our results since the filter is disconnected.
-    cy.findAllByTestId("cell-data")
-      .should("contain", "christ")
-      .and("contain", "xavier")
-      .and("not.contain", "kale");
-
-    // Reloading the dashboard breaks the filter in the original issue
-    cy.reload();
-
-    cy.log(
-      "Make sure the UI shows the filter is not connected to the GUI card",
-    );
-    editDashboard();
-
-    cy.findByTestId("fixed-width-filters").icon("gear").click();
-    getDashboardCard().should("contain", "Unknown Field");
-
-    cy.get("@dashboardId").then(id => {
-      cy.log("Make sure this works for the public sharing setting");
-      cy.request("POST", `/api/dashboard/${id}/public_link`).then(
-        ({ body: { uuid } }) => {
-          // Set the filter through the URL
-          cy.visit(`/public/dashboard/${uuid}?equal_to=3`);
-        },
-      );
-      cy.findAllByTestId("cell-data")
-        .should("contain", "christ")
-        .and("contain", "xavier")
-        .and("not.contain", "kale");
-
-      cy.log("Make sure this works for the embedding setting");
-      cy.request("PUT", `/api/dashboard/${id}`, {
-        embedding_params: {
-          equal_to: "locked",
-        },
-        enable_embedding: true,
-      });
-
-      const payload = {
-        resource: { dashboard: id },
-        params: { equal_to: [3] },
-      };
-
-      visitEmbeddedPage(payload);
-      cy.findAllByTestId("cell-data")
-        .should("contain", "christ")
-        .and("contain", "xavier")
-        .and("not.contain", "kale");
-    });
-  });
 });
 
 function connectFilterToColumn(column) {
