+++ frontend/src/metabase/dashboard/components/DashCard/DashCardParameterMapper/DashCardCardParameterMapperButton.tsx
@@ -103,6 +103,7 @@ export const DashCardCardParameterMapperButton = ({
           buttonText: t`Unknown Field`,
           buttonIcon: (
             <CloseIconButton
+              aria-label={t`Disconnect`}
               onClick={e => {
                 handleChangeTarget(null);
                 e.stopPropagation();
